refactor: move things of libs folder that are used just in one place to theclosest place that they are called
