+++ src/libs/fetcher.ts
@@ -1,14 +1,32 @@
 import { parseCookies } from 'nookies'
-import { Either, left, right } from 'fp-ts/Either'
+import { Either, left, right, isRight } from 'fp-ts/Either'
+import { pipe } from 'fp-ts/function'
+import { PathReporter } from 'io-ts/PathReporter'
 import {
   DefaultError,
   handleFetcherErrors,
   UnknownError,
-  validateCodec,
+  DecodeError,
 } from '@/libs'
 import * as t from 'io-ts'
 
 const baseURL = process.env.NEXT_PUBLIC_API_BASE_URL
+
+const validationHandler = <A>(
+  dataValidation: t.Validation<A>
+): Either<DecodeError, A> => {
+  if (isRight(dataValidation)) {
+    return right(dataValidation.right)
+  } else {
+    const decodeErrors = PathReporter.report(dataValidation)
+    return left(new DecodeError(decodeErrors.join(', ')))
+  }
+}
+
+export const validateCodec = <A>(
+  codec: t.Type<A, unknown, unknown>,
+  data: A
+): Either<DecodeError, A> => pipe(data, codec.decode, validationHandler<A>)
 
 export const fetcher = async <D, A>(
   path: string,
