+++ src/libs/index.ts
@@ -3,4 +3,3 @@ export * from '@/libs/cookies'
 export * from '@/libs/errors'
 export * from '@/libs/expression'
 export * from '@/libs/fetcher'
-export * from '@/libs/validate-codec'
